# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/me/profile": {
      "patch": {
        "operationId": "MeController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ]
      }
    },
    "/me/roles/choose": {
      "post": {
        "operationId": "MeController_chooseRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChooseRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ]
      }
    },
    "/me/roles/switch": {
      "post": {
        "operationId": "MeController_switchRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ]
      }
    },
...
    "/me/sessions/{id}": {
      "delete": {
        "operationId": "MeController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "me"
        ]
      }
    },
...
    "/staff/auth/logout": {
      "post": {
        "operationId": "StaffAuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-auth"
        ]
      }
    },
...
    "/me/contracts/{contractId}/payment": {
      "post": {
        "operationId": "PaymentController_create",
        "parameters": [
          {
            "name": "contractId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me-payments"
        ]
      }
    },
    "/me/payments": {
      "get": {
        "operationId": "MePaymentController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "SUCCEEDED",
                "FAILED",
                "CANCELLED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "contractId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me-payments"
        ]
      }
    },
    "/me/payments/{id}": {
      "get": {
        "operationId": "MePaymentController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me-payments"
        ]
      }
    },
...
    "/staff/payments/{id}": {
      "get": {
        "operationId": "StaffPaymentController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPaymentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bobo&Doda API",
    "description": "Markaziy Osiyo freelance marketplace — milestone escrow, double-entry ledger. Frontend `lib/api` chegarasini qoplaydi.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+998901234567"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+998901234567"
          },
          "code": {
            "type": "string",
            "example": "482913",
            "description": "6 xonali OTP kodi"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "AuthSessionDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Qisqa umrli JWT — `Authorization: Bearer <token>`"
          },
          "activeRole": {
            "type": "string",
            "enum": [
              "SELLER",
              "BUYER"
            ],
            "nullable": true,
            "description": "NULL — rol hali tanlanmagan"
          },
          "roleChosen": {
            "type": "boolean"
          },
          "profileDone": {
            "type": "boolean"
          },
          "isNewUser": {
            "type": "boolean"
          }
        },
        "required": [
          "accessToken",
          "roleChosen",
          "profileDone",
          "isNewUser"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "nullable": true
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "avatarKey": {
            "type": "object",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "SELLER",
                "BUYER"
              ]
            }
          },
          "activeRole": {
            "type": "string",
            "enum": [
              "SELLER",
              "BUYER"
            ],
            "nullable": true
          },
          "roleChosen": {
            "type": "boolean"
          },
          "profileDone": {
            "type": "boolean"
          },
          "verified": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "BLOCKED"
            ]
          },
          "sellerStatus": {
            "type": "string",
            "enum": [
              "NOT_APPLIED",
              "PENDING",
              "APPROVED",
              "REJECTED",
              "SUSPENDED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "phone",
          "roles",
          "roleChosen",
          "profileDone",
          "verified",
          "status",
          "sellerStatus",
          "createdAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          }
        },
        "required": [
          "fullName"
        ]
      },
      "ChooseRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "SELLER",
              "BUYER"
            ],
            "example": "BUYER"
          }
        },
        "required": [
          "role"
        ]
      },
      "SwitchRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "SELLER",
              "BUYER"
            ],
            "example": "SELLER"
          }
        },
        "required": [
          "role"
        ]
      },
...
      "PaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "failureReason": {
            "type": "object"
          },
          "succeededAt": {
            "type": "object"
          },
          "failedAt": {
            "type": "object"
          },
          "cancelledAt": {
            "type": "object"
          },
          "expiresAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "contractId",
          "provider",
          "status",
          "amount",
          "currency",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffPaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "failureReason": {
            "type": "object"
          },
          "succeededAt": {
            "type": "object"
          },
          "failedAt": {
            "type": "object"
          },
          "cancelledAt": {
            "type": "object"
          },
          "expiresAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "payerUserId": {
            "type": "string"
          },
          "providerPaymentId": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "contractId",
          "provider",
          "status",
          "amount",
          "currency",
          "createdAt",
          "updatedAt",
          "payerUserId"
        ]
      },
...
    }
  }
}